Tableau: Superstore
Imported data from Superstore, organized data, and utilized information on the store's products, marketplaces and dates to create visualizations. These visualizations create strong arguments on which products the store should focus on selling and possibly discontinue, certain marketplaces in the US with higher or lower sales/returns, and times of the year when sales are highest.